pluto: use uint64_t for byte-counts in SAs (kernel uses __u64)

+++ programs/pluto/kernel_netlink.c
@@ -2282,7 +2282,7 @@ add_entry:
  * @param sa Kernel SA to be queried
  * @return bool True if successful
  */
-static bool netlink_get_sa(const struct kernel_sa *sa, u_int *bytes,
+static bool netlink_get_sa(const struct kernel_sa *sa, uint64_t *bytes,
 		uint64_t *add_time)
 {
 	struct {
@@ -2313,11 +2313,7 @@ static bool netlink_get_sa(const struct kernel_sa *sa, u_int *bytes,
 				sa->text_said))
 		return FALSE;
 
-	/*
-	 * ??? CLANG 3.5 think that "Assigned value is garbage or undefined".
-	 * It's right.  See <linux/xfrm.h>: bytes has type __u64.
-	 */
-	*bytes = (u_int) rsp.info.curlft.bytes;
+	*bytes = rsp.info.curlft.bytes;
 	*add_time = rsp.info.curlft.add_time;
 	return TRUE;
 }
